refactor(color-map): change file name to match class name
