+++ lib/models/ui/linear_segmented_hsv_colormap.dart
